+++ stories/text-field.stories.js
@@ -1,15 +1,17 @@
-import React from 'react'
-
-import { storiesOf } from '@storybook/react'
+import React, { useState } from 'react'
 import { withA11y } from '@storybook/addon-a11y'
 import { withKnobs, text, number, boolean, select } from '@storybook/addon-knobs'
 
 import { TextField } from '../src/components/index.ts'
 
-storiesOf('Components/Forms/TextField', module)
-  .addDecorator(withA11y)
-  .addDecorator(withKnobs)
-  .add('normal state', () => (
+export default {
+  title: 'Components/Forms/TextField',
+  component: TextField,
+  decorators: [withA11y, withKnobs],
+}
+
+export const NormalState = () => {
+  return (
     <>
       {boolean('show label', false) ? <label>Input:</label> : null}
       <TextField placeholder="This is an input" />
@@ -22,26 +24,54 @@ storiesOf('Components/Forms/TextField', module)
       ) : null}
       <TextField rowCount={3} placeholder="This is an textarea" />
     </>
-  ))
-  .add('highlighted state', () => (
+  )
+}
+
+export const HighlightedState = () => {
+  return (
     <>
       <TextField placeholder="This is a highlighted input" isHighlighted />
       <TextField rowCount={3} placeholder="This is a highlighted textarea" isHighlighted />
     </>
-  ))
-  .add('error state', () => (
+  )
+}
+
+export const ErrorState = () => {
+  const [errors, setErrors] = useState({
+    customError: { customData: 18298 }
+  })
+
+  return (
     <>
-      <TextField placeholder="This is a error input" dataType="number" isRequired />
-      <TextField rowCount={3} placeholder="This is a error textarea" dataType="email" />
+      <TextField placeholder="This is a error input" dataType="number" hideErrors={boolean('hide errors', false)} isRequired />
+      <TextField rowCount={3} placeholder="This is a error textarea" dataType="email" hideErrors={boolean('hide errors', false)} />
+      <h3>With custom error</h3>
+      <pre>
+        {JSON.stringify(errors, null, 4)}
+      </pre>
+      <TextField
+        placeholder="This is a error input"
+        dataType="number"
+        errors={errors}
+        hideErrors={boolean('hide errors', false)}
+        onErrors={err => setErrors(err)}
+        isRequired
+      />
     </>
-  ))
-  .add('disabled state', () => (
+  )
+}
+
+export const DisabledState = () => {
+  return (
     <>
       <TextField placeholder="This is a disabled input" isDisabled />
       <TextField rowCount={3} placeholder="This is a disabled textarea" isDisabled />
     </>
-  ))
-  .add('full example', () => (
+  )
+}
+
+export const FullExample = () => {
+  return (
     <TextField
       placeholder={text('placeholder', 'Fill the field')}
       dataType={select('data type', ['text', 'email', 'number'])}
@@ -49,5 +79,7 @@ storiesOf('Components/Forms/TextField', module)
       isHighlighted={boolean('highlighted', false)}
       isDisabled={boolean('disabled', false)}
       isSmall={boolean('is small', false)}
+      hideErrors={boolean('hide errors', false)}
     />
-  ))
+  )
+}